This is Python set 3 Exercise modified

+++ Exrecise 3/Q.No.4.py
@@ -1,4 +1,4 @@
-year = float(input("enter year:"))
+year = int(input("enter year:"))
 
 if year % 4 == 0:
     if year % 100 == 0:
